+++ ContactDetailsApi/V1/Domain/Sns/ContactDetailsSns.cs
@@ -4,7 +4,7 @@ namespace ContactDetailsApi.V1.Domain.Sns
 {
     public class ContactDetailsSns
     {
-        public string Id { get; set; }
+        public Guid Id { get; set; }
         public string EventType { get; set; }
         public string SourceDomain { get; set; }
         public string SourceSystem { get; set; }
@@ -12,6 +12,6 @@ namespace ContactDetailsApi.V1.Domain.Sns
         public string CorrelationId { get; set; }
         public DateTime DateTime { get; set; }
         public User User { get; set; }
-        public string EntityId { get; set; }
+        public Guid EntityId { get; set; }
     }
 }
