+++ ContactDetailsApi/V1/Domain/Sns/User.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace ContactDetailsApi.V1.Domain.Sns
 {
     public class User
     {
-        public string Id { get; set; }
+        public Guid Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
     }
