+++ ContactDetailsApi/V1/Factories/EntityFactory.cs
@@ -4,55 +4,10 @@ using ContactDetailsApi.V1.Infrastructure;
 using System.Collections.Generic;
 using System.Linq;
 using ContactDetailsApi.V1.Boundary.Request;
-using Hackney.Core.JWT;
+using User = ContactDetailsApi.V1.Domain.User;
 
 namespace ContactDetailsApi.V1.Factories
 {
-    public class ContactDetailsSnsFactory : ISnsFactory
-    {
-        public ContactDetailsSns Create(ContactDetailsRequestObject contactDetails, Token token)
-        {
-            return new ContactDetailsSns
-            {
-                CorrelationId = Guid.NewGuid(),
-                DateTime = DateTime.UtcNow,
-                EntityId = contactDetails.Id,
-                Id = Guid.NewGuid(),
-                EventType = CreateEventConstants.EVENTTYPE,
-                Version = CreateEventConstants.V1VERSION,
-                SourceDomain = CreateEventConstants.SOURCEDOMAIN,
-                SourceSystem = CreateEventConstants.SOURCESYSTEM,
-                User = new User { Id = Guid.NewGuid(), Name = token.Name, Email = token.Email }
-            };
-        }
-    }
-
-    public interface ISnsFactory
-    {
-        ContactDetailsSns Create(ContactDetailsRequestObject contactDetails, Token token);
-    }
-
-    public class ContactDetailsSns
-    {
-        public Guid Id { get; set; }
-
-        public string EventType { get; set; }
-
-        public string SourceDomain { get; set; }
-
-        public string SourceSystem { get; set; }
-
-        public string Version { get; set; }
-
-        public Guid CorrelationId { get; set; }
-
-        public DateTime DateTime { get; set; }
-
-        public User User { get; set; }
-
-        public Guid EntityId { get; set; }
-    }
-
     public static class EntityFactory
     {
         public static ContactDetails ToDomain(this ContactDetailsEntity databaseEntity)
