+++ ContactDetailsApi/V1/Gateways/ISnsGateway.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using ContactDetailsApi.V1.Domain.Sns;
 using ContactDetailsApi.V1.Factories;
 
 namespace ContactDetailsApi.V1.Gateways
